Removed bad macro usage

+++ utilities.hpp
@@ -18,15 +18,15 @@ namespace std {
  template<> struct hash<sf::Vector2i> {
     public:
         std::size_t operator()(const sf::Vector2i &a)const {
-            return CityHash64((char *) &a, sizeof(a));
+            return static_cast<size_t>(CityHash64((char *) &a, sizeof(a)));
         }
     };
 }
 inline void textureRectangleToVertex(sf::IntRect a, sf::Vertex* vert){
-	vert[0].texCoords = sf::Vector2f(a.left, a.top);
-	vert[1].texCoords = sf::Vector2f(a.left + a.width, a.top);
-	vert[2].texCoords = sf::Vector2f(a.left + a.width, a.top + a.height);
-	vert[3].texCoords = sf::Vector2f(a.left, a.top + a.height);
+	vert[0].texCoords = sf::Vector2f(float(a.left), float(a.top));
+    vert[1].texCoords = sf::Vector2f(float(a.left + a.width), float( a.top));
+    vert[2].texCoords = sf::Vector2f(float(a.left + a.width), float(a.top + a.height));
+	vert[3].texCoords = sf::Vector2f(float(a.left),float( a.top + a.height));
 }
 inline sf::Color getFromJson(Json::Value v){
     return sf::Color(v.get("R", 0).asInt(), v.get("G", 0).asInt(), v.get("B", 0).asInt());
